backend/requirements.txt
Django==5.2.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
openai==1.12.0
PyPDF2==3.0.1
python-dotenv==1.0.0


